backend/requirements.txt
fastapi
uvicorn
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
langchain
langchain-google-genai
langgraph
python-dotenv
pydantic
httpx
python-multipart